moving upload dialog to its own file
